Return PDF bytes and make generator importable
